Refactor server fact loading into a separate class
